move passwords to dedicated packages
